Fixed operator new argument type

+++ lib/cm2/include/cm2/game.hpp
@@ -40,7 +40,7 @@ public:
 
 	CM2Game& operator=(const CM2Game& other);
 
-	static void* __cdecl operator new(unsigned int, FObject*, const FName&, const FName&, unsigned int);
+	static void* __cdecl operator new(size_t, FObject*, const FName&, const FName&, unsigned int);
 
 	int Init() override;
 	void Exit() override;
