+++ lib/faktum/include/faktum/game.hpp
@@ -21,7 +21,7 @@ public:
 
 	FGame& operator=(const FGame& other);
 
-	static void* __cdecl operator new(unsigned int, FObject*, const FName&, const FName&, unsigned int);
+	static void* __cdecl operator new(size_t, FObject*, const FName&, const FName&, unsigned int);
 
 	void DrawLevel(FLevel* level, FCanvas* canvas, FScene::TYPE sceneType);
 	FEditor* GetEditor() const;
